Initial commit of split repository.
After 10 hours of various attempts with git filter-branch, looping manually
through git rev-list, and other shenanigans, I've given up on trying to preserve
the history. The history is in https://github.com/zendframework/zend-expressive,
but at this point, I have better things to do than try and preserve it here.

So, here we are, with the Router subcomponent for zend-expressive, in its stable
glory pre-1.0.