Re-instate route result observer, and add subject
This patch re-instates (de-deprecates) the
`RouteResultObserverInterface`, but does so by adding its complement,
the `RouteResultSubjectInterface`, which notifies observers of
calculated route results. This latter interface is to be used with
`Zend\Expressive\Application` (which already implements these methods
on its master branch), but also allows observers the ability to depend
only on this package for purposes of testing.