RouterInterface tweaks
Removed the RuntimeConfigurableRouterInterface and put addRoute back into
the RouterInterface. We'll allow the `$options` to be passed on to the
actual routers via the new third `$options` parameter.

This approach does break BC as well, but it will be released with a new
major version to avoid surprising library users.