Added implicit(Head|Options) methods to Route
In order to determine if a matched route supports HEAD and/or OPTIONS
requests implicitly (i.e., the methods were not specified during
construction), this patch adds the methods `implicitHead()` and
`implicitOptions()`.

This will allow dispatchers to determine if they should delegate to
matched middleware, or instead provide automated responses (e.g.,
providing an empty 200 response with an `Allow` header for OPTIONS
requests; dispatching the middleware using `GET` and removing the
body for a `HEAD` request).

One step towards solving zendframework/zend-expressive#398.