Revert "Clarify that HEAD and OPTIONS should ALWAYS be allowed"
This reverts commit a81bf26435015cc28f2bc411460488e31ec2fd8f.

Per #28, auto-adding the HEAD and OPTIONS methods to each route causes
`Zend\Expressive\Application` to raise an exception; additionally, when
working on zendframework/zend-expressive-fastroute#24, I noticed that
tests broke due to duplicate paths having the same HTTP method defined;
adding HEAD and/or OPTIONS was causing adding routes with GET to result
in exceptions.

There are other ways for the routers to accommodate this, which can be
done with the `implicitHead()` and `implicitOptions()` methods of the
routes themselves, or by opportunistically checking for any method to
match against a given path.