Refactor routing and dispatch logic
This patch accomplishes several things.

First, it extracts the logic for producing a 405 response to new
middleware, `Zend\Expressive\Router\MethodNotAllowedMiddleware`. This
middleware checks for a route result that is due to method failure, and,
in such a case, uses the composed response prototype.

Second, the `RouteMiddleware` was modified as follows:

- it no longer requires a response prototype
- it _always_ injects the route result as a request attribute
- it no longer produces a 405 response on its own

Third, the `Route` class was modified to implement the PSR-15
`MiddlewareInterface`. When processed, it proxies to the composed
middleware instance.

Fourth, the `RouteResult` was modified to implement the PSR-15
`MiddlewareInterface`. When processed, if it is a route failure, it
delegates to the provided handler. If it is a successful route match, it
proxies to the matched `Route` instance. The `getMatchedMiddleware()`
method was removed, as it's essentially unnecessary.

Fifth, the `DispatchMiddleware` was modified to process the
`RouteResult` instance pulled from the request, instead of retrieving
its composed middleware.