Improve testing of ImplicitHeadMiddleware
Improves the testInvokesHandlerWhenRouteImplicitlySupportsHeadAndSupportsGet
logic to properly exercise a case where a HEAD request is issued, the
route does not explicitly support it, but _does_ support a GET request.

In doing so, discovered that the final return statement was incorrectly
using the response prototype, instead of the response returned from the
GET request.